analysis setting
